Working planner params and demo. Video included but crap
Video has the color channels shifted w.r.t. each other. No idea why and not going to investigate today. First 2:08 minutes of the video is with rviz open, then gazebo only (which makes it run much nicer)
Python code is still a bit of a mess but it works.