Charuco not relying on odom, xyzw copy
The tf tree is unusuable to do localization based on charuco because the charuco marker is not relative to the world, but relative to the turtlebot (which is relative to odom and the whole point is to not depend on odom).
By manually specifying the marker position, it's possible to create the required transformations to get the turtlebot position, relative to the marker, relative to world.

The various .x=.x, .y=.y, ... assignments have been replaced by general-purpose functions. In most cases this made the code much shorter.